requirements.txt
azure-core==1.30.2
azure-storage-blob==12.22.0
fastapi==0.111.0
fastapi-cli==0.0.7
httpcore==1.0.5
httptools==0.6.4
httpx==0.28.1
requests==2.32.3
selenium==4.23.1
cryptography==45.0.2
pymongo==4.13.0
mongoengine==0.29.1
pyodbc==5.2.0
python-dotenv==1.0.1
python-multipart==0.0.20
uvicorn==0.29.0
openpyxl==3.1.3
pandas==2.2.2
beautifulsoup4==4.12.3
validate-docbr==1.10.0
pydantic==2.7.1
pydantic_core==2.18.2